// src/enums.ts
export enum Align {
	FlexStart = 'FlexStart',
	Center = 'Center',
	FlexEnd = 'FlexEnd',
	Stretch = 'Stretch',
	// Baseline,
}

export enum AlignContent {
	FlexStart = 'FlexStart',
	Center = 'Center',
	FlexEnd = 'FlexEnd',
	SpaceAround = 'SpaceAround',
	SpaceBetween = 'SpaceBetween',
	Stretch = 'Stretch',
}

export enum Direction {
	Column = 'Column',
	ColumnReverse = 'ColumnReverse',
	Row = 'Row',
	RowReverse = 'RowReverse',
}

export enum JustifyContent {
	FlexStart = 'FlexStart',
	Center = 'Center',
	FlexEnd = 'FlexEnd',
	SpaceAround = 'SpaceAround',
	SpaceBetween = 'SpaceBetween',
	SpaceEvenly = 'SpaceEvenly',
}

export enum Wrap {
	NoWrap = 'NoWrap',
	Wrap = 'Wrap',
	WrapReverse = 'WrapReverse',
}